requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.1
filelock==3.20.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.22.2
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lightning==2.6.0
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.2
pyannote.audio==3.1.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.45
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.1.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.1.2
torchmetrics==1.8.2
tqdm==4.67.1
triton==2.1.0
typer==0.21.0
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
yarl==1.22.0

whisperx==3.2.0
faster-whisper==1.0.0